rename: 명확화를 위해 파일명 변경 Table -> View(#90)
