+++ src/features/game/components/ranking/team/TeamPitcherRankingView.tsx
@@ -1,22 +1,32 @@
-import { teamBatter as mockRanking } from '@/assets/data/__test__/mockRanking.json';
-import data from '@/assets/data/__test__/season-team-rank/seasonBatter.json';
-import { TeamRankingBatterConfig } from '@/constants/chart-config';
-import { teamBatterRankColumns } from '@/constants/columns/team-rank-colums';
+// import { TeamPitcherRank } from '@/features/common/types/pitchers';
+// import { useTeamRank } from "@/features/game/hooks/ranking/useTeamRank";
+import { teamPitcher as mockRanking } from '@/assets/data/__test__/mockRanking.json';
+import data from '@/assets/data/__test__/season-team-rank/seasonPitcher.json';
+import { TeamRankingPitcherConfig } from '@/constants/chart-config';
+import { teamPitcherRankColumns } from '@/constants/columns/team-rank-colums';
 import CustomBarChart from '@/features/common/CustomBarChart';
-// import { TeamBatterRank } from '@/features/common/types/batters';
-// import { useTeamRank } from '@/assets/hooks/ranking/useTeamRank';
 import DataTable from '@/features/common/DataTable';
 import { Config } from '@/features/player/components/PlayerRecordChart';
-import { SeasonBatter } from '@/features/player/types/record';
+import { SeasonPitcher } from '@/features/player/types/record';
 import { useState } from 'react';
 
-const seasonBatterData: SeasonBatter[] = data.data.list;
+const seasonPitcherData: SeasonPitcher[] = data.data.list;
 
-function TeamBatterRankingTable() {
+function TeamPitcherRankingView() {
   const [selectedTab, setSelectedTab] = useState<'table' | 'chart'>('table');
   const [chartConfig, setChartConfig] = useState<Config>(
-    TeamRankingBatterConfig
+    TeamRankingPitcherConfig
   );
+
+  // const { ranking, loading, error } = useTeamRank('pitcher');
+
+  // if (!ranking.length || loading) {
+  //   return null;
+  // }
+
+  // if (error) {
+  //   return <div>{error}</div>;
+  // }
 
   const handleConfig = (dataKey: keyof Config) => {
     const currentConfig = Object.keys(chartConfig).filter(
@@ -43,16 +53,6 @@ function TeamBatterRankingTable() {
     });
   };
 
-  // const { ranking, loading, error } = useTeamRank('batter');
-
-  // if (!ranking.length || loading) {
-  //   return null;
-  // }
-
-  // if (error) {
-  //   return <div>{error}</div>;
-  // }
-
   return (
     <div>
       <div className="flex items-center gap-2 justify-end">
@@ -78,13 +78,13 @@ function TeamBatterRankingTable() {
       {selectedTab === 'table' ? (
         <DataTable
           data={mockRanking}
-          columns={teamBatterRankColumns}
+          columns={teamPitcherRankColumns}
           domain="all"
         />
       ) : (
         <div>
           <CustomBarChart
-            data={seasonBatterData}
+            data={seasonPitcherData}
             config={chartConfig}
             XAxisKey={'teamName'}
           />
@@ -113,4 +113,4 @@ function TeamBatterRankingTable() {
   );
 }
 
-export { TeamBatterRankingTable };
+export { TeamPitcherRankingView };
